Add first run at desired chat functionality
This contains a forced way of checking for text. Does not include proper media, but allows for temporary handling of messages by both bots. It includes a system for sending various idle messages if they chat while in the middle of a mission. Cleans up debug information. Includes code for more robust progress handling, but will not be used for the first version.